+++ src/Utility/Reporter.php
@@ -27,38 +27,13 @@ class Reporter
 
         if (is_null($pid) || $pid === -1) {
             // Fork fails or there is no `pcntl` extension.
-            self::send($stats);
+            Request::post(self::ENDPOINT, $stats);
         } elseif ($pid === 0) {
             // Child process.
             posix_setsid();
-            self::send($stats);
+            Request::post(self::ENDPOINT, $stats);
             // Close child process after doing job.
             exit(0);
-        }
-    }
-
-    /**
-     * @param array $stats
-     */
-    private static function send(array $stats)
-    {
-        if (extension_loaded('curl')) {
-            $body = json_encode($stats, JSON_PRETTY_PRINT);
-            $ch = curl_init(self::ENDPOINT);
-            curl_setopt($ch, CURLOPT_HTTPHEADER, [
-                'Content-Type: application/json',
-                'Content-Length: ' . strlen($body)
-            ]);
-            curl_setopt($ch, CURLOPT_CUSTOMREQUEST, 'GET');
-            curl_setopt($ch, CURLOPT_POSTFIELDS, $body);
-            curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-            curl_setopt($ch, CURLOPT_FOLLOWLOCATION, true);
-            curl_setopt($ch, CURLOPT_MAXREDIRS, 10);
-            curl_setopt($ch, CURLOPT_CONNECTTIMEOUT, 5);
-            curl_setopt($ch, CURLOPT_TIMEOUT, 5);
-            curl_exec($ch);
-        } else {
-            file_get_contents(self::ENDPOINT . '?' . http_build_query($stats));
         }
     }
 }